+++ server/src/GameLogic/Bullets.cs
@@ -17,7 +17,7 @@ public interface IBullet
     public BulletType Type { get; }
 
     public Position BulletPosition { get; set; }
-    public double BulletSpeed { get; }
+    public float BulletSpeed { get; }
     public int BulletDamage { get; }
 
     public bool AntiArmor { get; }
@@ -28,25 +28,25 @@ public interface IBullet
 /// <summary>
 /// Default bullet for a weapon. Used by Cannon.
 /// </summary>
-public class Bullet(Position position, double speed, int damage, bool antiArmor = false) : IBullet
+public class Bullet(Position position, float speed, int damage, bool antiArmor = false) : IBullet
 {
     public IBullet.BulletType Type => IBullet.BulletType.Bullet;
 
     public Position BulletPosition { get; set; } = position;
 
-    public double BulletSpeed { get; } = speed;
+    public float BulletSpeed { get; } = speed;
     public int BulletDamage { get; } = damage;
 
     public bool AntiArmor { get; } = antiArmor;
 }
 
-public class LaserBullet(Position position, double speed, int damage, bool antiArmor = false) : IBullet
+public class LaserBullet(Position position, float speed, int damage, bool antiArmor = false) : IBullet
 {
     public IBullet.BulletType Type => IBullet.BulletType.LaserBullet;
 
     public Position BulletPosition { get; set; } = position;
 
-    public double BulletSpeed { get; } = speed;
+    public float BulletSpeed { get; } = speed;
     public int BulletDamage { get; } = damage;
 
     public bool AntiArmor { get; } = antiArmor;
