+++ server/src/GameLogic/Constants/Constants.cs
@@ -16,10 +16,10 @@ public static partial class Constants
     public const int KNIFE_REMAINING_TIME = 10;
 
     //Weapon
-    public const double INITIAL_ATTACK_SPEED = 0.1;
-    public const double ATTACK_SPEED_INCREASE_FACTOR = 2;
-    public const double INITIAL_BULLET_SPEED = 2;
-    public const double BULLET_SPEED_INCREASE_FACTOR = 2;
+    public const float INITIAL_ATTACK_SPEED = 0.1f;
+    public const float ATTACK_SPEED_INCREASE_FACTOR = 2;
+    public const float INITIAL_BULLET_SPEED = 2;
+    public const float BULLET_SPEED_INCREASE_FACTOR = 2;
     public const int INITIAL_DAMAGE = 1;
     public const int DAMAGE_INCREASE = 1;
     public const int INITIAL_BULLETS = 5;
@@ -30,17 +30,17 @@ public static partial class Constants
     public const float MOVE_SPEED = 0.1f;
     public const float TURN_SPEED = (float)Math.PI / 60;
     public const float PLAYER_RADIUS = 0.5f;
-    public const double BULLET_GENERATE_DISTANCE = 0.8;
+    public const float BULLET_GENERATE_DISTANCE = 0.8f;
 
     // Bullet
     public const float BULLET_RADIUS = 0.1f;
     public const int BULLET_REMAINING_TICKS = 100;
-    public const double LASER_WIDTH = 0.1;
-    public const double LASER_LENGTH_EQUAVALENT_BULLET_FLYING_TICKS = 20;
+    public const float LASER_WIDTH = 0.1f;
+    public const float LASER_LENGTH_EQUAVALENT_BULLET_FLYING_TICKS = 20;
 
     //Map
     public const float WALL_LENGTH = 10.0f;
-    public const double WALL_THICK = 0.1;
+    public const float WALL_THICK = 0.1f;
 
     // Wall
     public const int BREAKABLE_WALL_MAX_COLLIDE_COUNT = 3;
